# reports/openapi-baseline.json
{
  "components": {
    "schemas": {
...
      "CreateBillingApiKeyRequest": {
        "additionalProperties": false,
        "description": "Self-service API key creation \u2014 v1.2.4 audit P0-1 successor\nof the former ``POST /v1/infra/keys``.",
        "example": {
          "tier": "free"
        },
        "properties": {
          "tier": {
            "default": "free",
            "title": "Tier",
            "type": "string"
          }
        },
        "title": "CreateBillingApiKeyRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "description": "API key passed directly in the X-API-Key header (alternative to Bearer).",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "BearerAuth": {
        "description": "API key passed as Bearer token in the Authorization header.",
        "scheme": "bearer",
        "type": "http"
      },
      "X402Payment": {
        "description": "x402 payment proof for stateless authentication.",
        "in": "header",
        "name": "X-PAYMENT",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "A2A Commerce Gateway",
    "version": "1.2.9"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/billing/discounts": {
      "get": {
        "operationId": "get_volume_discount_v1_billing_discounts_get",
        "parameters": [
          {
            "in": "query",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tool_name",
            "required": true,
            "schema": {
              "title": "Tool Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "quantity",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Quantity",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Volume Discount",
        "tags": [
          "billing"
        ]
      }
    },
    "/v1/billing/estimate": {
      "get": {
        "operationId": "estimate_cost_v1_billing_estimate_get",
        "parameters": [
          {
            "in": "query",
            "name": "tool_name",
            "required": true,
            "schema": {
              "title": "Tool Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "quantity",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Quantity",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "agent_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Estimate Cost",
        "tags": [
          "billing"
        ]
      }
    },
    "/v1/billing/exchange-rates": {
      "get": {
        "operationId": "get_exchange_rate_v1_billing_exchange_rates_get",
        "parameters": [
          {
            "in": "query",
            "name": "from_currency",
            "required": true,
            "schema": {
              "title": "From Currency",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to_currency",
            "required": true,
            "schema": {
              "title": "To Currency",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Exchange Rate",
        "tags": [
          "billing"
        ]
      }
    },
    "/v1/billing/keys": {
      "get": {
        "description": "List API keys owned by the authenticated caller.\n\nv1.2.4 audit P0-1: this is the non-admin self-service path.\n``create_api_key`` is reused as the ``require_tool`` anchor\n(it is NOT in ``ADMIN_ONLY_TOOLS``) so free/pro tiers reach\nthis handler without hitting the admin gate. The handler then\ncalls the ``_list_api_keys`` implementation directly, always\nscoped to the authenticated caller's ``agent_id`` via ownership.",
        "operationId": "list_billing_api_keys_v1_billing_keys_get",
        "parameters": [
          {
            "in": "query",
            "name": "agent_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Billing Api Keys",
        "tags": [
          "billing"
        ]
      },
      "post": {
        "operationId": "create_billing_api_key_v1_billing_keys_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBillingApiKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Billing Api Key",
        "tags": [
          "billing"
        ]
      }
    },
    "/v1/billing/leaderboard": {
      "get": {
        "operationId": "get_agent_leaderboard_v1_billing_leaderboard_get",
        "parameters": [
          {
            "in": "query",
            "name": "metric",
            "required": false,
            "schema": {
              "default": "spend",
              "title": "Metric",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Agent Leaderboard",
        "tags": [
          "billing"
        ]
      }
    },
    "/v1/billing/wallets": {
      "post": {
        "operationId": "create_wallet_v1_billing_wallets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWalletRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Wallet",
        "tags": [
          "billing"
        ]
      }
    },
    "/v1/billing/wallets/{agent_id}/analytics": {
      "get": {
        "operationId": "get_service_analytics_v1_billing_wallets__agent_id__analytics_get",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Service Analytics",
        "tags": [
          "billing"
        ]
      }
    },
    "/v1/billing/wallets/{agent_id}/balance": {
      "get": {
        "operationId": "get_balance_v1_billing_wallets__agent_id__balance_get",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "currency",
            "required": false,
            "schema": {
              "default": "CREDITS",
              "title": "Currency",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Balance",
        "tags": [
          "billing"
        ]
      }
    },
    "/v1/billing/wallets/{agent_id}/budget": {
      "get": {
        "operationId": "get_budget_status_v1_billing_wallets__agent_id__budget_get",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Budget Status",
        "tags": [
          "billing"
        ]
      },
      "put": {
        "operationId": "set_budget_cap_v1_billing_wallets__agent_id__budget_put",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BudgetCapRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Budget Cap",
        "tags": [
          "billing"
        ]
      }
    },
    "/v1/billing/wallets/{agent_id}/convert": {
      "post": {
        "operationId": "convert_currency_v1_billing_wallets__agent_id__convert_post",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConvertCurrencyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Convert Currency",
        "tags": [
          "billing"
        ]
      }
    },
    "/v1/billing/wallets/{agent_id}/deposit": {
      "post": {
        "operationId": "deposit_v1_billing_wallets__agent_id__deposit_post",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Deposit",
        "tags": [
          "billing"
        ]
      }
    },
    "/v1/billing/wallets/{agent_id}/freeze": {
      "post": {
        "operationId": "freeze_wallet_v1_billing_wallets__agent_id__freeze_post",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Freeze Wallet",
        "tags": [
          "billing"
        ]
      }
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    },
    {
      "ApiKeyAuth": []
    },
    {
      "X402Payment": []
    }
  ]
}